refactor: keep the window size change store next to its only consumer
Reverts the move to src/hooks and the width only filter. The store keeps the
orientation check, and its header now says why the flag lives in a module
store rather than in each screen.
